Fix [object Object] is not a future
When ios-sim-portable is used as dependency of any module that has fibers, the fibers in ios-sim-portable conflict at runtime with the original package's fibers.
So add postinstall script to check if there are fibers in the upper level dir and upper level dir is called node_modules. If these circumstances are fulfilled, remove the ios-sim-portable's fibers directory - let's use original package's fibers module.