Add --sdk as parameter
When CLI uses ios-sim-portable through require("ios-sim-portable"), this module must understand the same command line parameters as the one passed to NativeScript CLI / AppBuilder CLI.
The option there is called `--sdk`, while here it is `--sdkVersion`. Add `--sdk` here, so ios-sim-portable will pass it and use it to start the correct emulator.