Add grunt pack command
Add grunt pack command that will:
* remove prepublish script from package.json
* clean the project (delete all .js files inside it)
* build .ts files in release mode
* call `npm pack` and produce .tgz for the new version

We have to remove the prepublish script as it does not allow us clean publish of the .tgz.
The problem is that when `npm publish <tgz>` is used, the package is added to local npm cache. After that it's package.json is extracted from the cached .tgz file and modified with additional metadata. At this point npm calls the prepublish script, but it fails to find our prepublish.js as it's not extracted in the cache (only the package.json is extracted).

I've not removed the prepublish script from the package.json, so if anyone uses `npm pack` locally or `npm publish` directly, the correct .js files will be used.

> NOTE: Calling `grunt pack` locally will modify your package.json. DO NOT commit these changes.