Fix detection of iPad Pro
When the device name contains `()`, the regex matches it instead of the GUID and after that we exclude the device from the list with available devices.
Such devices are:
```
    iPad Pro (9.7 inch) (7FF984D4-0755-432D-BE0E-6EB44F0489CB) (Shutdown)
    iPad Pro (12.9 inch) (F02012C8-6D4D-46FF-90D7-5DF90EF579E8) (Booted)
```

This way users are unable to execute any operation on such simulators.
Fix the regular expression to match based on the GUID instead of using the `(` as main part of the regex.

Check this gist for comparison of the results with the old and the new regex:
https://gist.github.com/rosen-vladimirov/85b4ccb78af2e604ac91efecd2a67150