fix: launchApplication fails when device identifier is passed
In case `launchApplication` is called with device identifier (instead of name), the validation fails as it expects a name. However, we already support starting and working with Id, so fix the validation.
Also, in the past there was only one instance of a simulator running. When you wanted to start application on another simulator, we were killing the existing one and starting the new one. This is no longer required as multiple iOS Simulators may be running simultaneously. So remove the logic for killing the simulator