fix: Do not wait 15 seconds to detect simulator
Currently when we want to detect iOS Simulators we have a setInterval that iterrates 30 times for 500ms.
When you do not have any iOS Simulator, this causes 15 seconds of waiting.
Just retry once and continue, if the simulators don't show up, well, they do not want to work with you.