fix: get_app_container returns non-existent path
In case application is not installed on simulator, `xcrun simctl get_app_container ...` command returns some random path that does not exist. In case you install the application after that, the path is not the same as the returned by the command before that.
So check if the path exists and return null in case it does not.