fix: getPid returns incorrect result
In case application name contains `grep`, the `getPid` method will not return any result as we have `grep -v grep` in the executed command.
Also, in case you run `tns run ios --path <path to app> --device <id>`, the `getPid` method will return the pid of CLI's process as well as it maches the criteria.

The fix is to check for `/${deviceId}/` when grep-ing. Also remove the `grep -v grep` and replace it with exclusion of the command we've executed. This is due to the fact that Node.js executes a new shell process and the command executed in `getPid` returns it as well.