fix for default simulator
this fix let the XCode simulator choose the default simulator to run. 
Which is the last state of the simulator app. (can even be multiple simulators)
Before that it was always choosing the same simulator. 
We now have the same behavior as the simulator app which brings a much better user experience ( imho)