fix: launch default simulator rather than iPhone 6 (#126)
* fix for default simulator

this fix let the XCode simulator choose the default simulator to run. 
Which is the last state of the simulator app. (can even be multiple simulators)
Before that it was always choosing the same simulator. 
We now have the same behavior as the simulator app which brings a much better user experience ( imho)

* chore: support no deviceId

* Update iphone-simulator-common.ts

* fix: don't check device state if it's null

Co-authored-by: Igor Randjelovic <rigor789@gmail.com>